refactor genre movie person

+++ src/main/java/com/sba301/online_ticket_sales/service/impl/GenreServiceImpl.java
@@ -23,6 +23,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Slf4j
@@ -32,14 +33,13 @@ public class GenreServiceImpl implements GenreService {
 
   GenreRepository genreRepository;
   GenreMapper genreMapper;
-  MovieRepository movieRepository;
 
   @Override
+  @Transactional
   public GenreResponse createGenre(GenreCreationRequest request) {
     try {
       Genre genre = genreMapper.toGenre(request);
       Genre savedGenre = genreRepository.save(genre);
-      log.info("Created genre: {}", savedGenre.getName());
       return genreMapper.toGenreResponse(savedGenre);
     } catch (DataIntegrityViolationException e) {
       throw new AppException(ErrorCode.GENRE_ALREADY_EXISTS);
@@ -47,13 +47,13 @@ public class GenreServiceImpl implements GenreService {
   }
 
   @Override
+  @Transactional
   public GenreResponse updateGenre(Integer id, GenreUpdateRequest request) {
     Genre genre =
         genreRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.GENRE_NOT_FOUND));
     try {
       genreMapper.updateGenreFromRequest(request, genre);
       Genre updatedGenre = genreRepository.save(genre);
-      log.info("Updated genre: {}", updatedGenre.getName());
       return genreMapper.toGenreResponse(updatedGenre);
     } catch (DataIntegrityViolationException e) {
       throw new AppException(ErrorCode.GENRE_ALREADY_EXISTS);
@@ -61,24 +61,17 @@ public class GenreServiceImpl implements GenreService {
   }
 
   @Override
+  @Transactional
   public void deleteGenre(Integer id) {
-    Genre genre =
-        genreRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.GENRE_NOT_FOUND));
 
-    // Tìm tất cả các Movie liên quan đến Genre này
-    List<Movie> movies = genreRepository.findMoviesByGenreId(id);
+    Genre genre = genreRepository.findById(id)
+            .orElseThrow(() -> new AppException(ErrorCode.GENRE_NOT_FOUND));
 
-    // Xóa Genre khỏi danh sách genres của các Movie
-    for (Movie movie : movies) {
-      movie.getGenres().remove(genre);
+    if (genreRepository.isGenreInUse(id)) {
+      throw new AppException(ErrorCode.GENRE_IN_USE);
     }
 
-    // Lưu lại các Movie đã cập nhật
-    movieRepository.saveAll(movies);
-
-    // Xóa Genre
     genreRepository.delete(genre);
-    log.info("Deleted genre: {}", genre.getName());
   }
 
   @Override
